Moved tabs into their own components
